fix: seal TramaiException, add ApprovalException, update docs, optional US_ASCII encoding
- Revert TramaiException from open class back to sealed class
- Add dev.tramai.core.exception.ApprovalException as direct subclass
- Move IllegalApprovalTransitionException to TramaiException.kt
  (dev.tramai.core.exception package), extends ApprovalException
- Delete old IllegalApprovalTransitionException.kt in approval package
- Update imports in InMemoryApprovalStore and test

- Update POLICY-API-DESIGN.md:
  - Add consumedBy/consumedAt to ApprovalRequest model
  - Add consumeApproved() to ApprovalStore SPI
  - Document constant-time token-digest comparison
  - Clarify PR #14 vs PR #15 responsibilities

- Use StandardCharsets.US_ASCII explicitly for digest byte conversion
  in consumeApproved() constant-time comparison

- Update PR description with accurate file/test counts
