+++ docs/security/POLICY-API-DESIGN.md
@@ -212,6 +212,8 @@ data class ApprovalRequest(
     val decidedBy: String?,
     val decidedAt: Instant?,
     val decisionComment: String?,
+    val consumedBy: String?,
+    val consumedAt: Instant?,
     val version: Long,
 )
 
@@ -225,6 +227,12 @@ interface ApprovalStore {
     suspend fun create(request: ApprovalRequest): ApprovalRequest
     suspend fun get(approvalId: String): ApprovalRequest?
     suspend fun transition(approvalId: String, expectedVersion: Long, transition: ApprovalTransition): ApprovalRequest
+    suspend fun consumeApproved(
+        approvalId: String,
+        expectedVersion: Long,
+        presentedTokenDigest: Sha256Digest,
+        consumedBy: String,
+    ): ApprovalRequest
 }
 ```
 
@@ -263,10 +271,13 @@ PENDING → APPROVED/DENIED/TIMED_OUT (all terminal). Timeout succeeds only when
 ### Key design decisions
 
 - Approval token (`approvalTokenDigest`) is a SHA-256 digest of a generated nonce. The raw token is provided to the requestor at creation time. PR #15 will consume and verify the raw token exactly once.
+- `consumeApproved()` uses constant-time `MessageDigest.isEqual()` with explicit `StandardCharsets.US_ASCII` encoding to prevent digest-comparison timing attacks.
+- `consumedBy` and `consumedAt` are recorded atomically with the version increment. Second consume of the same approval fails with clear message.
 - `decidedBy` is non-nullable on `Approve`/`Deny` transitions.
 - Comments are optional on transitions.
 - Version starts at 0 and is incremented atomically. No CAS retry loop — `ConcurrentHashMap.compute()` provides the atomicity.
 - No engine integration yet. PR #15 will build workflow suspension/resume on this foundation.
+- PR #14 scope: domain model, store SPI, validation, expiry, deterministic tests. PR #15 scope: workflow suspension, raw-token engine integration, engine-level resume, auto-deny job.
 
 ## Original Design (Phase 0) — Approval Request — Full Binding
